@@ -1,7 +1,8 @@
 {
   "name": "well-manager-react",
-  "version": "1.3.1",
+  "version": "1.3.2",
   "private": true,
+  "homepage": "/id2x/ClientApp/",
   "dependencies": {
     "@progress/kendo-data-query": "^1.5.5",
     "@progress/kendo-drawing": "^1.16.2",
@@ -17,77 +18,65 @@
     "@progress/kendo-react-indicators": "^5.0.1",
     "@progress/kendo-react-inputs": "^5.0.1",
     "@progress/kendo-react-intl": "^5.0.1",
+    "@progress/kendo-react-labels": "^5.3.0",
     "@progress/kendo-react-layout": "^5.0.1",
     "@progress/kendo-react-listbox": "^5.0.1",
     "@progress/kendo-react-notification": "^5.3.0",
+    "@progress/kendo-react-popup": "^5.5.0",
     "@progress/kendo-react-progressbars": "^5.0.1",
     "@progress/kendo-react-treeview": "^5.0.1",
     "@progress/kendo-theme-default": "^5.2.0",
-    "@reduxjs/toolkit": "^1.6.1",
+    "async": "^3.2.4",
+    "buffer": "^6.0.3",
     "bootstrap": "^4.1.3",
     "calculate-size": "^1.1.1",
     "canvas2svg": "^1.0.16",
-    "cldr-core": "^39.0.0",
-    "cldr-dates-full": "^39.0.0",
-    "cldr-numbers-full": "^39.0.0",
+    "cldr-core": "^41.0.0",
+    "cldr-dates-full": "^41.0.0",
+    "cldr-numbers-full": "^41.0.0",
     "flexlayout-react": "^0.5.14",
-    "i18next": "^20.4.0",
-    "jquery": "^3.5.1",
+    "i18next": "^21.8.11",
     "jspdf": "^2.5.1",
-    "konva": "^8.3.0",
     "lodash": "^4.17.21",
-    "node-xml-lite": "0.0.7",
-    "on-el-resize": "0.0.4",
     "parse-color": "^1.0.0",
-    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-i18next": "^11.11.4",
-    "react-konva": "^17.0.2-5",
-    "react-redux": "^7.2.5",
-    "react-router": "^6.3.0",
-    "react-router-bootstrap": "^0.25.0",
+    "react-i18next": "^11.17.2",
+    "react-redux": "^7.2.6",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "^3.4.4",
-    "reactstrap": "^8.4.1",
-    "recharts": "^2.1.9",
+    "react-scripts": "5.0.1",
+    "recharts": "^2.1.10",
     "redux": "^4.2.0",
-    "rimraf": "^2.6.2",
+    "redux-thunk": "^2.4.1",
+    "stream": "^0.0.2",
     "svg2pdf.js": "^2.2.0",
-    "three": "^0.139.2",
-    "three-trackballcontrols": "^0.9.0"
+    "web-vitals": "^2.1.4",
+    "xml-js": "^1.6.11"
   },
   "devDependencies": {
-    "@progress/kendo-react-labels": "^5.3.0",
-    "ajv": "^6.9.1",
-    "babel-polyfill": "^6.26.0",
-    "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^6.3.2",
-    "cross-env": "^5.2.0",
-    "dotenv-webpack": "^7.0.3",
-    "eslint": "^6.8.0",
-    "eslint-config-react-app": "^5.2.0",
-    "eslint-plugin-flowtype": "^4.6.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.18.3",
-    "nan": "^2.14.1",
-    "postcss": "^8.3.11",
-    "postcss-loader": "^4.3.0",
-    "ts-loader": "^9.3.0",
-    "typescript": "^4.6.4",
-    "webpack": "^4.42.0",
-    "webpack-cli": "^4.9.1"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.11.41",
+    "@types/parse-color": "^1.0.1",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
+    "cross-env": "^7.0.3",
+    "typescript": "^4.7.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "webpack",
-    "test": "cross-env CI=true react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "lint": "eslint ./src/"
+    "start": "cross-env GENERATE_SOURCEMAP=false react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
