@@ -1,19 +1,19 @@
-// module pixelPerMeter
+import { once } from "lodash";
 
-var _ = require( "lodash" )
 
-module.exports = _.once( () => {
-	var pxpcm = document.createElement( "div" )
-	var sampleSizeInCm = 20
+export default once( () => {
+	const div = document.createElement('div');
+	// размер "образца" в см
+	const sampleSizeInCm = 20;
 
-	pxpcm.style.width = pxpcm.style.height = sampleSizeInCm + "cm"
-	pxpcm.style.position = "fixed"
+	div.style.width = div.style.height = sampleSizeInCm + 'cm';
+	div.style.position = 'fixed';
+	document.body.appendChild(div);
 
-	document.body.appendChild( pxpcm )
 	try {
-		return pxpcm.clientWidth * ( 100 / sampleSizeInCm )
+		return div.clientWidth * (100 / sampleSizeInCm);
 	}
 	finally {
-		document.body.removeChild( pxpcm )
+		document.body.removeChild(div);
 	}
-} )
+});
