@@ -1,35 +1,85 @@
-import React from 'react';
-import { Provider } from 'react-redux';
-import SessionLoader from './SessionLoader';
-import createSessionManager from '../dataManagers/SessionManager';
-import store from '../store/store';
-import {load} from "@progress/kendo-react-intl";
-import likelySubtags from "cldr-core/supplemental/likelySubtags.json";
-import currencyData from "cldr-core/supplemental/currencyData.json";
-import weekData from "cldr-core/supplemental/weekData.json";
-import numbers from "cldr-numbers-full/main/ru/numbers.json";
-import caGregorian from "cldr-dates-full/main/ru/ca-gregorian.json";
-import dateFields from "cldr-dates-full/main/ru/dateFields.json";
-import timeZoneNames from "cldr-dates-full/main/ru/timeZoneNames.json";
+import React, { useEffect, useMemo } from "react";
+import { Link, useParams } from "react-router-dom";
+import { useTranslation } from "react-i18next";
+import { useDispatch, useSelector } from "react-redux";
+import { setSystemName } from "../store/actionCreators/appState";
+import { startSession } from "../store/thunks";
 
-load(
-    likelySubtags,
-    currencyData,
-    weekData,
-    numbers,
-    caGregorian,
-    dateFields,
-    timeZoneNames
-);
+import LoadingStatus from "./common/LoadingStatus";
+import SessionLoader from "./SessionLoader";
 
 
-export default function SystemRoot(props) {
-    const { systemName } = props;
-    createSessionManager(systemName, store);
-
-    return (
-        <Provider store={store}>
-          <SessionLoader />
-        </Provider>
-    );
+/** Проверяет, есть ли указанная система в списке доступных систем. */
+const checkSystem = (systemName, systemList) => {
+  return systemList.find((system) => system.id === systemName);
 }
+
+/** Компонент, использующийся если система не найдена. */
+const SystemNotFound = ({root, name}) => {
+  const { t } = useTranslation();
+  return (
+    <div style={{padding: '2em', fontSize: '1.15rem'}}>
+      <div>{t('systems.notFound', {systemName: name})}</div>
+      <Link to={root}>&#11176; {t('systems.backToSystemList')}</Link>
+    </div>
+  );
+}
+
+/** Корень системы. Route: `/systems/:systemID`. */
+const SystemRoot = ({root}) => {
+  const dispatch = useDispatch();
+  const { systemID } = useParams();
+
+  const appState = useSelector((state) => state.appState);
+  const sessionManager = useSelector((state) => state.sessionManager);
+
+  const isSystemExist = useMemo(() => {
+    return appState.systemList.data && checkSystem(appState.systemID, appState.systemList.data);
+  }, [appState]);
+
+  const readyToStartSession = useMemo(() => {
+    return appState.systemList.loaded && !appState.sessionID.loaded && appState.systemID && sessionManager;
+  }, [appState, sessionManager])
+
+  const allReady = useMemo(() => {
+    return appState.systemID
+      && appState.config.loaded && appState.config.success
+      && appState.systemList.loaded && appState.systemList.success
+      && appState.sessionID.loaded && appState.sessionID.success;
+  }, [appState]);
+
+  useEffect(() => {
+    if (systemID !== appState.systemID) {
+      dispatch(setSystemName(systemID));
+    }
+    if (readyToStartSession) {
+      dispatch(startSession(sessionManager.startSession));
+    }
+  }, [dispatch, systemID, appState, sessionManager, readyToStartSession]);
+
+  if (allReady) {
+    return isSystemExist ? <SessionLoader /> : <SystemNotFound name={systemID} root={root}/>;
+  }
+
+  if (!appState.config.loaded) {
+    return <LoadingStatus loadingType={'config'}/>;
+  }
+  if (appState.config.success === false) {
+    return <LoadingStatus loadingType={'config'} success={false}/>;
+  }
+  if (!appState.systemList.loaded) {
+    return <LoadingStatus loadingType={'systems'}/>;
+  }
+  if (appState.systemList.success === false) {
+    return <LoadingStatus loadingType={'systems'} success={false}/>;
+  }
+  if (!appState.sessionID.loaded) {
+    return <LoadingStatus loadingType={'session'} />;
+  }
+  if (appState.sessionID.success === false) {
+    return <LoadingStatus loadingType={'session'} success={false}/>
+  }
+  return <LoadingStatus loadingType={''}/>;
+}
+
+export default SystemRoot;
