@@ -1,15 +1,4 @@
-// module cp1251
-
-var map = [].map
-
-export function toUnicode( str) {
-	return map.call( str, x => {
-		var c = x & 255
-		return String.fromCharCode( c < 128 ? c : table1251[ c - 128 ] )
-	} ).join( "" )
-}
-
-var table1251 = [
+const table1251 = [
 	0x0402, 0x0403, 0x201A, 0x0453, 0x201E, 0x2026, 0x2020, 0x2021, 0x20AC, 0x2030, 0x0409,
 	0x2039, 0x040A, 0x040C, 0x040B, 0x040F, 0x0452, 0x2018, 0x2019, 0x201C, 0x201D, 0x2022,
 	0x2013, 0x2014, 0x0020, 0x2122, 0x0459, 0x203A, 0x045A, 0x045C, 0x045B, 0x045F, 0x00A0,
@@ -22,4 +11,11 @@ var table1251 = [
 	0x0433, 0x0434, 0x0435, 0x0436, 0x0437, 0x0438, 0x0439, 0x043A, 0x043B, 0x043C, 0x043D,
 	0x043E, 0x043F, 0x0440, 0x0441, 0x0442, 0x0443, 0x0444, 0x0445, 0x0446, 0x0447, 0x0448,
 	0x0449, 0x044A, 0x044B, 0x044C, 0x044D, 0x044E, 0x044F
-]
+];
+
+export function toUnicode( str) {
+	return [].map.call(str, x => {
+		const c = x & 255;
+		return String.fromCharCode(c < 128 ? c : table1251[c - 128]);
+	}).join('');
+}
