@@ -1,19 +1,15 @@
-﻿import React from 'react';
-import FormHeader from './Form/FormHeader';
+﻿import React from "react";
+import FormHeader from "./Form/FormHeader";
+import { formIconsDict } from "../dicts/images";
 
-export default function Screenshot(props) {
-    const { formData } = props;
 
-    const getImagePath = (formType) => {
-        return window.location.pathname + 'images/' + formType + '.PNG';
-    }
-
-    return (
-        <div className="screenshot-container">
-            <FormHeader formData={formData} />
-            <div className="imgbox">
-                <img src={getImagePath(formData.type)} alt="logo" />
-            </div>
-        </div>
-    );
+export default function Screenshot({formData}) {
+  return (
+    <div className={'screenshot-container'}>
+      <FormHeader formData={formData} />
+      <div className={'imgbox'}>
+        <img src={formIconsDict[formData.type]} alt={formData.type} />
+      </div>
+    </div>
+  );
 }
