@@ -1,7 +1,8 @@
-﻿import {createStore} from "redux";
+﻿import {createStore , applyMiddleware} from "redux";
+import thunk from "redux-thunk";
 import reducer from "./rootReducer";
 
 
-const store = createStore(reducer);
+const store = createStore(reducer, applyMiddleware(thunk));
 
 export default store;
